#35 refactor: user - 모듈명 수정
- 유저 모듈명 변경
