+++ settings.gradle
@@ -1,3 +1,3 @@
 rootProject.name = 'fifteen-night-backend-develop'
 
-include 'common', 'gateway', 'server', 'product-service', 'company-service', 'hub-service', 'user' , 'delivery-service'
+include 'common', 'gateway', 'server', 'company-service', 'product-service','hub-service', 'user-service' , 'delivery-service'
