+++ user-service/src/main/java/com/fn/eureka/client/userservice/application/dto/auth/request/UserSignInRequestDto.java
@@ -1,0 +1,11 @@
+package com.fn.eureka.client.userservice.application.dto.auth.request;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class UserSignInRequestDto {
+	private String userName;
+	private String userPassword;
+}
