+++ user-service/src/main/java/com/fn/eureka/client/userservice/application/dto/auth/response/UserSignInResponseDto.java
@@ -1,0 +1,11 @@
+package com.fn.eureka.client.userservice.application.dto.auth.response;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class UserSignInResponseDto {
+	private String accessToken;
+	private String refreshToken;
+}
