+++ user-service/src/main/java/com/fn/eureka/client/userservice/application/dto/user/response/UserUpdateResponseDto.java
@@ -1,0 +1,18 @@
+package com.fn.eureka.client.userservice.application.dto.user.response;
+
+import java.util.UUID;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class UserUpdateResponseDto {
+	private UUID userId;
+	private String userName;
+	private String userNickname;
+	private String userEmail;
+	private String userRole;
+	private String userPhone;
+	private String userSlackId;
+}
