+++ user-service/src/test/java/com/fn/eureka/client/userservice/UserApplicationTests.java
@@ -1,0 +1,13 @@
+package com.fn.eureka.client.userservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class UserApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
