--- user/src/main/java/com/fn/eureka/client/user/UserApplication.java
@@ -1,15 +1,0 @@
-package com.fn.eureka.client.user;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@SpringBootApplication
-@EnableJpaAuditing
-public class UserApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(UserApplication.class, args);
-	}
-
-}
