--- user/src/main/resources/application.yml
@@ -1,37 +1,0 @@
-spring:
-  application:
-    name: user-service
-  config:
-    import:
-      - file:user/.env[.properties]
-#      - classpath:./properties/datasource.yml
-
-  datasource:
-    url: ${DB_URL} # 환경변수로 설정
-    username: ${DB_USER}  # 환경변수에서 불러오기
-    password: ${DB_PASSWORD}
-    driver-class-name: org.postgresql.Driver
-
-  jpa:
-    database-platform: org.hibernate.dialect.PostgreSQLDialect
-    hibernate:
-      ddl-auto: update  # 테이블 자동 생성 (개발용)
-    show-sql: true  # SQL 로그 출력
-    properties:
-      hibernate:
-        format_sql: true  # SQL 보기 좋게 출력
-
-service:
-  jwt:
-    access-expiration: ${JWT_ACCESS_EXPIRATION}
-    refresh-expiration: ${JWT_REFRESH_EXPIRATION}
-    secret-key: ${JWT_SECRET_KEY}
-
-
-server:
-  port: ${SERVER_PORT}  # 서버 포트 환경변수로 설정
-
-eureka:
-  client:
-    service-url:
-      defaultZone: ${EUREKA_SERVER_URL}  # 환경변수로 설정
