--- user/src/test/java/com/fn/eureka/client/user/UserApplicationTests.java
@@ -1,13 +1,0 @@
-package com.fn.eureka.client.user;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class UserApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
